Split the AMD/ROCm guards out of cuda_runtime into their own module
cuda_runtime.py had drifted into carrying the AMD present-hardware checks,
the ROCm fail-loud guard, and now the rocBLAS coverage check, until half
the module contradicted its name. The AMD half moves to rocm_runtime.py,
the vendor-neutral ldd and probe-diagnostic helpers to
engine_diagnostics.py, and the two identical soname-matching functions
collapse into one links_any. The assert_gpu_devices_usable dispatcher is
gone: planning calls the CUDA and ROCm asserts directly, each self-gating
the way the CUDA one always did. Tests split along the same lines and
every mock target follows its symbol.
