+++ src/lilbee/providers/fleet/cuda_runtime.py
@@ -17,14 +17,13 @@ from __future__ import annotations
 import importlib.util
 import logging
 import os
-import shutil
-import subprocess
 import sys
 from pathlib import Path
 from typing import TYPE_CHECKING
 
 from lilbee.providers import model_cache
 from lilbee.providers.base import ProviderError
+from lilbee.providers.fleet.engine_diagnostics import device_probe_diagnostic, links_any
 
 if TYPE_CHECKING:
     from lilbee.providers.fleet.devices import FleetDevice
@@ -40,24 +39,8 @@ _CUDA_WHEEL_IMPORTS: tuple[str, ...] = (
 # The sonames those wheels provide, used to tell from ``ldd`` whether a binary is a
 # CUDA build (it lists the soname whether or not the runtime resolves).
 _CUDA_SONAMES: tuple[str, ...] = ("libcudart.so.12", "libcublas.so.12", "libnvrtc.so.12")
-# The HIP equivalents. A ROCm build links these and none of the CUDA sonames, so
-# the CUDA guard above never fired for it.
-_HIP_SONAMES: tuple[str, ...] = ("libamdhip64.so", "librocblas.so", "libhipblas.so")
-# Substrings that mark a CUDA init failure in the engine's --list-devices output.
-_CUDA_ERROR_MARKERS: tuple[str, ...] = ("error", "fail", "no cuda")
-_LDD_TIMEOUT_S = 10
-# PCI vendor id for AMD, as sysfs reports it.
-log = logging.getLogger(__name__)
 
-_AMD_PCI_VENDOR_ID = "0x1002"
-# How much of the probe output to quote when no specific error line is found.
-_DIAGNOSTIC_TAIL_CHARS = 300
-# Backends whose devices run rocBLAS; ggml prints either name depending on version.
-_AMD_BACKENDS = ("ROCm", "HIP")
-# AMD's own escape hatch: makes the runtime treat every device as the given
-# gfx version, which is how a near-miss card (gfx1031) runs a shipped
-# architecture's kernels (gfx1030).
-_HSA_OVERRIDE_VAR = "HSA_OVERRIDE_GFX_VERSION"
+log = logging.getLogger(__name__)
 
 
 def _wheel_lib_dir(import_name: str) -> Path | None:
@@ -91,12 +74,10 @@ def cuda_runtime_env() -> dict[str, str]:
     if not dirs:
         return {}
     parts = [str(d) for d in dirs]
-    # Drop existing entries that are already wheel dirs so calling this on every
-    # reload pass (apply_cuda_runtime_env) is idempotent instead of accumulating
-    # duplicate copies that get baked into each spawned server's environment.
-    wheel_dirs = set(parts)
     existing = os.environ.get("LD_LIBRARY_PATH", "")
-    parts.extend(entry for entry in existing.split(os.pathsep) if entry and entry not in wheel_dirs)
+    for part in existing.split(os.pathsep):
+        if part and part not in parts:
+            parts.append(part)
     return {"LD_LIBRARY_PATH": os.pathsep.join(parts)}
 
 
@@ -110,250 +91,9 @@ def apply_cuda_runtime_env() -> None:
     os.environ.update(cuda_runtime_env())
 
 
-def _ldd_output(binary: Path, env: dict[str, str]) -> str | None:
-    """``ldd`` stdout for *binary* under *env*; None when ldd can't run on it."""
-    ldd = shutil.which("ldd")
-    if ldd is None:
-        return None
-    try:
-        proc = subprocess.run(  # noqa: S603 - ldd path and the resolved binary
-            [ldd, str(binary)],
-            capture_output=True,
-            text=True,
-            timeout=_LDD_TIMEOUT_S,
-            env=env,
-            check=False,
-        )
-    except (OSError, subprocess.SubprocessError):
-        # Not an ELF, a static binary, or a timeout: nothing to inspect.
-        return None
-    return proc.stdout
-
-
 def _links_cuda_runtime(binary: Path, env: dict[str, str]) -> bool:
     """True when *binary* lists a CUDA runtime soname (a CUDA build), resolved or not."""
-    out = _ldd_output(binary, env)
-    if out is None:
-        return False
-    return any(soname in out for soname in _CUDA_SONAMES)
-
-
-def _device_probe_diagnostic(probe_output: str) -> str:
-    """The probe's CUDA error line, or a short tail of its output."""
-    out = probe_output.strip()
-    for line in out.splitlines():
-        lowered = line.lower()
-        if "cuda" in lowered and any(marker in lowered for marker in _CUDA_ERROR_MARKERS):
-            return line.strip()
-    return out[-_DIAGNOSTIC_TAIL_CHARS:] if out else "(the engine's device probe printed nothing)"
-
-
-def _links_hip_runtime(binary: Path, env: dict[str, str]) -> bool:
-    """True when *binary* lists a HIP runtime soname (a ROCm build), resolved or not."""
-    out = _ldd_output(binary, env)
-    if out is None:
-        return False
-    return any(soname in out for soname in _HIP_SONAMES)
-
-
-def _amd_gpu_present() -> bool:
-    """Whether the kernel exposes an AMD GPU, without needing ROCm to work.
-
-    Read from sysfs rather than from amd-smi or rocm-smi: those ship with ROCm,
-    and the failure this guards is precisely ROCm being installed wrong, so a
-    tool-based check would report "no GPU" for the case it exists to catch.
-    """
-    if not Path("/dev/kfd").exists():
-        return False
-    for vendor in Path("/sys/class/drm").glob("card*/device/vendor"):
-        try:
-            if vendor.read_text().strip().lower() == _AMD_PCI_VENDOR_ID:
-                return True
-        except OSError:
-            continue
-    return False
-
-
-def _amd_discrete_gpu_proven() -> bool:
-    """Whether a discrete AMD card is positively known to be present.
-
-    Positive evidence only. The sysfs checks that find an AMD GPU cannot tell a
-    discrete card from an APU, and the difference decides between refusing to
-    start and merely running slower, so the question is put to the Vulkan loader,
-    which reports the device type. An unreachable loader proves nothing and
-    answers no, which keeps the softer path.
-    """
-    from lilbee.providers.fleet.gpu_select import PCIVendorID, discrete_gpu_from_vendor
-
-    return discrete_gpu_from_vendor(PCIVendorID.AMD) is True
-
-
-def _gfx_name(target_version: int) -> str:
-    """The gfx name for a KFD ``gfx_target_version`` (90010 is gfx90a: minor and
-    step print as hex)."""
-    major, rest = divmod(target_version, 10000)
-    minor, step = divmod(rest, 100)
-    return f"gfx{major}{minor:x}{step:x}"
-
-
-def _host_amd_gfx_targets() -> set[str]:
-    """The gfx targets of this host's AMD GPUs, from the kernel driver's KFD topology.
-
-    Read from sysfs for the same reason as :func:`_amd_gpu_present`: ROCm tooling
-    is part of what can be broken. CPU nodes report a target version of 0. Empty
-    when the topology is absent or unreadable, which is "no claim".
-    """
-    targets: set[str] = set()
-    for props in Path("/sys/class/kfd/kfd/topology/nodes").glob("*/properties"):
-        try:
-            text = props.read_text()
-        except OSError:
-            continue
-        for line in text.splitlines():
-            name, _, value = line.partition(" ")
-            if name == "gfx_target_version" and value.strip().isdigit() and int(value.strip()):
-                targets.add(_gfx_name(int(value.strip())))
-    return targets
-
-
-def _bundled_rocblas_gfx_targets(binary: Path) -> set[str] | None:
-    """The gfx targets the rocBLAS kernels bundled beside *binary* cover.
-
-    rocBLAS loads a ``TensileLibrary_lazy_<gfx>.dat`` for the running card and
-    aborts the process when there is none, so the shipped masters are the ground
-    truth for which cards this build can run a GEMM on. ``None`` when no bundle
-    sits beside the binary (a system-ROCm engine): that is "no claim", where an
-    empty set would mean "supports nothing".
-    """
-    library = binary.parent / "rocblas" / "library"
-    if not library.is_dir():
-        return None
-    return {
-        f.name.removeprefix("TensileLibrary_lazy_").removesuffix(".dat")
-        for f in library.glob("TensileLibrary_lazy_gfx*.dat")
-    }
-
-
-def _hsa_override_gfx() -> str | None:
-    """The gfx target a user-set ``HSA_OVERRIDE_GFX_VERSION`` maps every device to."""
-    raw = os.environ.get(_HSA_OVERRIDE_VAR, "")
-    try:
-        major, minor, step = (int(part) for part in raw.split("."))
-    except ValueError:
-        return None
-    return _gfx_name(major * 10000 + minor * 100 + step)
-
-
-def _rocm_support_facts(binary: Path) -> str:
-    """What is known about shipped kernels and host gfx targets, as message text."""
-    shipped = _bundled_rocblas_gfx_targets(binary)
-    parts = []
-    if shipped:
-        parts.append(f"This build ships GPU kernels for: {', '.join(sorted(shipped))}.")
-    if host := _host_amd_gfx_targets():
-        parts.append(f"This host's AMD GPU targets: {', '.join(sorted(host))}.")
-    return " " + " ".join(parts) if parts else ""
-
-
-def _assert_rocblas_covers_enumerated_devices(binary: Path, devices: list[FleetDevice]) -> None:
-    """Refuse a card the bundle ships no rocBLAS kernels for, before rocBLAS aborts.
-
-    An enumerated device is no proof of support: the engine's device code and
-    rocBLAS's kernels are built from different lists, and a card covered by the
-    first but not the second initializes fine and then aborts the whole engine at
-    the first batched matrix multiply. Support is read from the bundled Tensile
-    masters rather than kept as a constant, so it cannot drift from what shipped.
-    """
-    if not any(d.backend in _AMD_BACKENDS for d in devices):
-        return
-    shipped = _bundled_rocblas_gfx_targets(binary)
-    if shipped is None:
-        return
-    if (override := _hsa_override_gfx()) is not None:
-        if override not in shipped:
-            log.warning(
-                "%s maps every AMD device to %s, but this build ships GPU kernels only "
-                "for: %s. The engine will abort at the first matrix multiplication if a "
-                "model runs on the GPU.",
-                _HSA_OVERRIDE_VAR,
-                override,
-                ", ".join(sorted(shipped)),
-            )
-        return
-    host = _host_amd_gfx_targets()
-    if not host:
-        return
-    missing = host - shipped
-    if not missing:
-        return
-    if host & shipped:
-        log.warning(
-            "This host has AMD GPU(s) with target %s, which this build ships no GPU "
-            "kernels for; the engine will abort if a model is placed on one. Restrict "
-            "HIP_VISIBLE_DEVICES to the supported cards, or set %s if the card is a "
-            "near miss of a shipped target (gfx1031 runs gfx1030 kernels with "
-            "%s=10.3.0).",
-            ", ".join(sorted(missing)),
-            _HSA_OVERRIDE_VAR,
-            _HSA_OVERRIDE_VAR,
-        )
-        return
-    raise ProviderError(
-        f"This host's AMD GPU is {', '.join(sorted(host))}, but this engine build ships "
-        f"GPU kernels only for: {', '.join(sorted(shipped))}. The engine would start and "
-        "then abort at the first matrix multiplication, so it is refused up front.\n"
-        f"If the card is a near miss of a shipped target, set {_HSA_OVERRIDE_VAR} to that "
-        f"target's version (a gfx1031 card runs the gfx1030 kernels with "
-        f"{_HSA_OVERRIDE_VAR}=10.3.0). Otherwise install lilbee's Vulkan build, which "
-        "supports AMD cards ROCm does not."
-    )
-
-
-def assert_gpu_devices_usable(binary: Path, devices: list[FleetDevice], probe_output: str) -> None:
-    """Fail loud when a GPU build cannot initialize any device on a host that has one.
-
-    Dispatches on what the binary actually links, so a ROCm build gets the same
-    treatment a CUDA build has always had. ROCm has a wide silent-failure class
-    (kernel and user-space version mismatch, an unsupported gfx target, no
-    permission on /dev/kfd) and every one of them previously ended as a quiet
-    fall back to CPU.
-    """
-    assert_cuda_devices_usable(binary, devices, probe_output)
-    if not sys.platform.startswith("linux"):
-        return
-    _assert_rocblas_covers_enumerated_devices(binary, devices)
-    if devices:
-        return
-    env = {**os.environ, **cuda_runtime_env()}
-    if not (_links_hip_runtime(binary, env) and _amd_gpu_present()):
-        return
-    if not _amd_discrete_gpu_proven():
-        # An APU is an AMD GPU by every check above: amdgpu exposes /dev/kfd for
-        # integrated parts too, and the iGPU carries vendor 0x1002. But AMD's
-        # population of GPUs ROCm legitimately does not support is large, and an
-        # unsupported gfx target is the normal case for an APU rather than a
-        # misconfiguration. Failing here would stop the engine on a laptop where
-        # CPU serving worked, which is worse than the slow fallback this guard
-        # exists to catch, so say so and let it start.
-        log.warning(
-            "The engine links the ROCm/HIP runtime and this host has an AMD GPU, but it "
-            "enumerated no device, so GPU work will fall back to CPU. No discrete AMD card "
-            "was found, so this is most likely an APU whose gfx target this ROCm build does "
-            "not support (check with 'rocminfo').%s The engine reported: %s",
-            _rocm_support_facts(binary),
-            _device_probe_diagnostic(probe_output),
-        )
-        return
-    raise ProviderError(
-        "The engine links the ROCm/HIP runtime and this host has an AMD GPU, but it "
-        "enumerated no device, so GPU work would silently fall back to CPU.\n"
-        f"The engine reported: {_device_probe_diagnostic(probe_output)}\n"
-        "Likely causes: the ROCm user-space version does not match the amdgpu kernel "
-        "driver; the GPU's gfx target is not supported by this ROCm build (check with "
-        "'rocminfo'); no read/write permission on /dev/kfd (the user is usually added "
-        "to the 'render' and 'video' groups); or a restrictive ROCR_VISIBLE_DEVICES or "
-        f"HIP_VISIBLE_DEVICES.{_rocm_support_facts(binary)}"
-    )
+    return links_any(binary, env, _CUDA_SONAMES)
 
 
 def assert_cuda_devices_usable(binary: Path, devices: list[FleetDevice], probe_output: str) -> None:
@@ -374,7 +114,7 @@ def assert_cuda_devices_usable(binary: Path, devices: list[FleetDevice], probe_o
         return
     if not model_cache.has_nvidia_gpu():
         return
-    diagnostic = _device_probe_diagnostic(probe_output)
+    diagnostic = device_probe_diagnostic(probe_output)
     raise ProviderError(
         "The engine links the CUDA runtime and this host has an NVIDIA GPU, but it "
         "enumerated no CUDA-capable device, so GPU work would silently fall back to CPU.\n"
